refactoring mock data for point
